Added Tabs
Es gibt noch en paar Probleme  ...  und zwat muss nach dem update der
Datenbank die App erst neu gestartet werden damit sich die screens
ändern ...
und auch das der stand der derdatenbank im subtitle ist hab ich wohl
gelöscht aber sonst funktionierts mal
