--- app/src/main/java/huti/sportinfo/GameDetails.java
@@ -1,10 +1,0 @@
-package huti.sportinfo;
-
-import android.support.v7.app.ActionBarActivity;
-
-/**
- * Created by Adrian on 08.01.2015.
- */
-public class GameDetails extends ActionBarActivity {
-
-}
