+++ app/src/main/java/huti/sportinfo/modules/ModuleFussball.java
@@ -4,6 +4,7 @@ import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
+import android.support.v7.app.ActionBarActivity;
 import android.text.Html;
 
 import java.text.ParseException;
@@ -18,7 +19,7 @@ import huti.sportinfo.SqliteHelper;
  */
 public class ModuleFussball {
 
-    private MainActivity activity = null;
+    private ActionBarActivity activity = null;
     private String url = "";
     private String kennung = "";
     private int inturlart = 0;
@@ -26,7 +27,7 @@ public class ModuleFussball {
     private int intsportart = 0;
     private int intlast = 0;
 
-    public ModuleFussball(MainActivity activity, String url, String kennung, int urlart, int idfavorit, int intsportart, int intlast) {
+    public ModuleFussball(ActionBarActivity activity, String url, String kennung, int urlart, int idfavorit, int intsportart, int intlast) {
         this.activity = activity;
         this.url = url;
         this.kennung = kennung;
