+++ app/src/main/res/values/arrays.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string-array name="planets_array" />
+    <string-array name="navigation_drawer_items" />
+</resources>
